型定義を修正

+++ extension_src/src/board_state.hpp
@@ -9,7 +9,7 @@
 using namespace godot;
 
 struct Cell {
-    int8_t type;
+    int type;
     int8_t side;
     bool is_promoted;
